[CIR][ABI] Create target lowering library skeleton
This patch adds a new TargetLowering library that intends to add supoort
for lowering CIR code to target specific CIR code. It is largely based
on the original codegen library used to lower AST nodes to LLVM IR
instructions. Because of this, each file has a comment specifying the
original codegen file that inspired the new file.

In some cases CIRGen ignores or partially hadles the ABI/target specific
codegen. This new library will be responsible for tackling these cases
where the generated code is not fully target specific. Some examples of
this are ctor/dtors and calling conventions, which are not fully handled
by CIRGen.

Some files are empty but will eventually be used and a few getters and
methods where added to avoid unused warnings. Missing features in this
library are tracked in a dedicated MissingFeature.h header.