[CIR][ABI] Implement call conv lowering for void CallOps
This patch implements the lowering of function calls that receive and
return void. In pratice, nothing has to be done (at least for the
x86 ABI), so this case is used as a primer for the target lowering
library since it helps populating the base logic for handling calling
convention lowering of function calls.